Ref animations. Disabled for desktop
